Break option parsing tests into multiple smaller tests
This breaks the option parsing tests into multiple smaller tests.
Doing this improves readability and makes it clearer what is being
tested, as well as what the outcomes should be for each test.

The coverage is increased to 100% for the option-parsing method of
the code, although it was very close already.

A new dev requirement of pytest-catchlog has been introduced to
ensure that the deprecation warning message is being shown. There are
other areas in which logging testing could be introduced, not only
for this one place (which wouldn't be worthwhile). If at a later
point no more logging tests have been added then it may be worth
revisiting if having the additional requirement is beneficial or not.