Improve test coverage for tesseract.
This improves the test coverage for tesseract module. It's not quite
up to 100% as the keyboard interrupt is hard to test.

Some functions have been broken up in the tesseract module to
facilitate testing, which has given additional methods available
as part of the public API, but the functionality of existing
methods has not been changed.