Switch to using packaging.version.parse instead of pkg_resources.parse_version
This is slightly arbitrary as both seem to work equally well, however as both
are external it makes sense to go with packaging as being a smaller and more
specific package rather than pulling pkg_resources out of setuptools.
Setuptools was also not listed as a specific dependency before, which was
bad.

Also a unit-test was added to ensure compatability with the new tesseract
`4.00.00alpha` version number added. The actual tesseract 4 binary has
not been tested for compatability.